+++ $shared/settings.ts
@@ -155,7 +155,9 @@ export type ConfigurationSettings = {
 	validate: Validate;
 	packageManager: PackageManagers;
 	useESLintClass: boolean;
-	experimentalUseFlatConfig: boolean;
+	experimental: {
+		useFlatConfig: boolean;
+	}
 	codeAction: CodeActionSettings;
 	codeActionOnSave: CodeActionsOnSaveSettings;
 	format: boolean;