+++ client/src/client.ts
@@ -605,7 +605,9 @@ export namespace ESLintClient {
 					validate: Validate.off,
 					packageManager: config.get<PackageManagers>('packageManager', 'npm'),
 					useESLintClass: config.get<boolean>('useESLintClass', false),
-					experimentalUseFlatConfig: config.get<boolean>('experimentalUseFlatConfig', false),
+					experimental: {
+						useFlatConfig: config.get<boolean>('experimental.useFlatConfig', false)
+					},
 					codeActionOnSave: {
 						mode: CodeActionsOnSaveMode.all
 					},