+++ playgrounds/flatConfig/eslint.config.js
@@ -1,4 +1,6 @@
 const globals = require('globals');
+const typescriptParser =  require('@typescript-eslint/parser');
+const typescriptPlugin = require('@typescript-eslint/eslint-plugin');
 
 module.exports = [
 	"eslint:recommended",
@@ -21,6 +23,31 @@ module.exports = [
 		rules: {
 			"no-undef": "warn",
 			"no-console": "warn"
+		}
+	},
+	{
+		files: ["*.ts", "**/*.ts"],
+		plugins: {
+			"@typescript-eslint": typescriptPlugin
+		},
+		languageOptions: {
+			parser: typescriptParser,
+			parserOptions: {
+				project: "./tsconfig.json",
+				sourceType: "module",
+				ecmaVersion: 2020
+			}
+		},
+		rules: {
+			"semi": "off",
+			"@typescript-eslint/semi": "error",
+			"no-extra-semi": "warn",
+			"curly": "warn",
+			"quotes": ["error", "single", { "allowTemplateLiterals": true } ],
+			"eqeqeq": "error",
+			"indent": "off",
+			"@typescript-eslint/indent": ["warn", "tab", { "SwitchCase": 1 } ],
+			"@typescript-eslint/no-floating-promises": "error"
 		}
 	}
 ]