+++ playgrounds/flatConfig/test.ts
@@ -1,0 +1,7 @@
+export function main(): number {
+	let s: string = '';
+	if (s == "") {
+		return 0;
+	}
+	return -1;
+}